Update script for contact generation with dynamic count
Expanded the contact generation function in 'create-contact.js' to utilize a dynamic count of contacts defined by environment variable 'TOTAL_CONTACT'. Previously, the script was hard-coded to create only 10 contacts irrespective of load requirements. With this change, it is possible to simulate different levels of loads by simply changing the 'TOTAL_CONTACT' value.